# spark-ai-server/src/main/resources/openapi/dynamic-data-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SPARK Server API",
    "version": "1.0.0",
    "description": "Complete SPARK server API documentation, including authentication, app config, page config, AI, projects, navigation, cache, diagnostics, metadata-driven dynamic tables, unified CRUD, DataView query, tree operations, batch jobs, and SSE notifications."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/ai/sessions": {
      "post": {
        "tags": [
          "AI Sessions"
        ],
        "summary": "Create AI session",
        "requestBody": {
          "$ref": "#/components/requestBodies/GenericJson"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericJson"
          }
        }
      },
      "delete": {
        "tags": [
          "AI Sessions"
        ],
        "summary": "Delete all AI sessions",
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          }
        }
      }
    },
    "/ai/sessions/{sessionId}/turn": {
      "post": {
        "tags": [
          "AI Sessions"
        ],
        "summary": "Run one AI session turn",
        "parameters": [
          {
            "$ref": "#/components/parameters/SessionId"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/GenericJson"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericJson"
          }
        }
      }
    },
    "/ai/sessions/{sessionId}/turn/stream": {
      "post": {
        "tags": [
          "AI Sessions"
        ],
        "summary": "Stream one AI session turn",
        "parameters": [
          {
            "$ref": "#/components/parameters/SessionId"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/GenericJson"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/SseStream"
          }
        }
      }
    },
    "/ai/sessions/{sessionId}/turn/append": {
      "post": {
        "tags": [
          "AI Sessions"
        ],
        "summary": "Append a turn result to an AI session",
        "parameters": [
          {
            "$ref": "#/components/parameters/SessionId"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/GenericJson"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericJson"
          }
        }
      }
    },
    "/ai/sessions/{sessionId}/append": {
      "post": {
        "tags": [
          "AI Sessions"
        ],
        "summary": "Append messages to an AI session",
        "parameters": [
          {
            "$ref": "#/components/parameters/SessionId"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/GenericJson"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericJson"
          }
        }
      }
    },
    "/ai/sessions/{sessionId}/conversation": {
      "get": {
        "tags": [
          "AI Sessions"
        ],
        "summary": "Get AI session conversation",
        "parameters": [
          {
            "$ref": "#/components/parameters/SessionId"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericJson"
          }
        }
      }
    },
    "/ai/sessions/{sessionId}": {
      "delete": {
        "tags": [
          "AI Sessions"
        ],
        "summary": "Delete one AI session",
        "parameters": [
          {
            "$ref": "#/components/parameters/SessionId"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
...
      "SessionId": {
        "name": "sessionId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "example": "session-001"
      },
...
    },
    "requestBodies": {
      "GenericJson": {
        "required": false,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/GenericObject"
            }
          }
        }
      },
...
    },
    "responses": {
      "GenericJson": {
        "description": "JSON response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/GenericObject"
            }
          }
        }
      },
      "SseStream": {
        "description": "Server-sent event stream",
        "content": {
          "text/event-stream": {
            "schema": {
              "type": "string"
            }
          }
        }
      },
...
      "Ok": {
        "description": "OK",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "ok": {
                  "type": "boolean",
                  "example": true
                }
              }
            }
          }
        }
      },
...
    },
    "schemas": {
      "GenericObject": {
        "type": "object",
        "additionalProperties": true
      },
...
    },
...
  }
}